requirements.txt
ircrobots    ~=0.5.0
pyyaml       ~=5.4.1
aiosqlite    ~=0.17.0
